Update kik_kre_2_0.md (#29)
Fixed up the notation and adjusted the KRE Carryover pool to not differentiate between different tracks. This way, if the DAS track is not fully utilized early in the year, the carryover pool will be used proportionally to the different tracks based on the allocation at a future time (and not what it was when the underutilization happened).