Kin Improvement Proposal V3 (#86)
* Kin Improvement Proposal for V3

This is an improvement proposal to the current KRE. It aims to simplify the KRE calculations while also improving the hold and buy tracks. The proposal has different headlines to the proposal template, as I thought some points needed a bit of elaboration.

* Add files via upload

This is an improvement proposal to the current KRE. It aims to simplify the KRE calculations while also improving the hold and buy tracks. The proposal has different headlines to the proposal template, as I thought some points needed a bit of elaboration.

* Update kik_kre_3.md

* Update kik_kre_3.md

* Update kik_kre_3.md

* Update kik_kre_3.md

* Update kik_kre_3.md

* Update kik_kre_3.md

* Update kik_kre_3.md

* Update kik_kre_3.md

* Update kik_kre_3.md

* Update kik_kre_3.md

* Update kik_kre_3.md

* Update kik_kre_3.md

* Rename kik_kre_3.md to kre_3.md