Accepted by KF
The second requirement has been accepted by the KF, for immediate effect. (Summarized below):

The user/developer are paid Kin at a rate at most the market rate for Kin as calculated on the date of purchase (i.e. a user can earn the developer at most $0.01 worth of Kin for an ad generating $0.01 of revenue, and a user buying $1.00 worth of kin cannot receive more than $1.00 worth of Kin).